chore: refactoring getting metadata from server
