+++ lib/ui/widgets/basewidget.js
@@ -1,9 +1,9 @@
 var React = require('react');
 var schema = require('@signalk/signalk-schema');
 var signalkClient = require('@signalk/client').Client;
-//var signalkUnits = new signalkClient;
+var signalkUnits = new signalkClient;
 
-//import {getConversionsForUnit} from '../settings/conversions'
+import {getConversionsForUnit} from '../settings/conversions'
 
 for (var key in schema.metadata) {
   schema.metadata[key].regex = new RegExp(key.replace('.', '\.').replace('*', '.*').replace('$', ''));
@@ -80,6 +80,49 @@ BaseWidget.prototype.getHandledSources = function() {
     key: this.options.key,
     sourceId: this.options.sourceId
   }];
+}
+
+BaseWidget.prototype.updateFromOnlineMeta = function(widget) {
+  widget.updateFromOnlineMeta = 1;
+  signalkUnits = new signalkClient(widget.instrumentPanel.host.split(':')[0], widget.instrumentPanel.host.split(':')[1], widget.instrumentPanel.useSSL);
+  signalkUnits.apiGet('/vessels/self/' + widget.options.path.replace(/\./g, '/') + '/meta/units')
+  .then(function (response) {
+//    console.log('updateFromOnlineMeta:'+response.status+':'+response.req.url+':'+response.text);
+    var newUnit = JSON.parse(response.text);
+    if ((newUnit !== '') &&
+        (widget.unit !== newUnit) &&
+        (widget.updateFromOnlineMeta === 1) &&
+        (widget.widgets.length > 0)) {
+      widget.unit = newUnit;
+      var valueStream = widget.streamBundle.getStreamForSourcePath(widget.options.sourceId, widget.options.path);
+      if (widget.unit === 'rad') {
+        widget.unit = 'deg';
+      }
+      if (widget.options.convertTo) {
+        var conversions = getConversionsForUnit(widget.unit);
+        if (conversions[widget.options.convertTo]) {
+          valueStream = valueStream.map(conversions[widget.options.convertTo]);
+        } else {
+          console.error("No such conversion:" + widget.unit + "=>" + widget.options.convertTo)
+        }
+      }
+//      console.log('updateFromOnlineMeta:'+widget.unit+':'+widget.options.path);
+      var newWidgets  = [];
+      widget.widgets.forEach(function(item, index) {
+        newWidgets.push( React.cloneElement(item,{
+          unit: widget.unit,
+          valueStream: valueStream
+        }));
+      })
+      widget.widgets = newWidgets;
+      widget.updateFromOnlineMeta = 0;
+      widget.instrumentPanel.pushGridChanges();
+    }
+  })
+  .catch(function (error) {
+    console.warn('updateFromOnlineMeta:'+error.status+':'+widget.options.path);
+    widget.updateFromOnlineSchema = 0;
+  });
 }
 
 module.exports = BaseWidget;
