+++ lib/ui/widgets/compass.js
@@ -40,7 +40,6 @@ require('util').inherits(Compass, BaseWidget);
 
 function Compass(id, options, streamBundle, instrumentPanel) {
   BaseWidget.call(this, id, options, streamBundle, instrumentPanel);
-  this.options = options;
   this.options.initialDimensions = this.options.initialDimensions || {w: 2, h: 4};
 
   var valueStream = streamBundle.getStreamForSourcePath(options.sourceId, options.path);
