+++ lib/ui/widgets/universal.js
@@ -9,8 +9,6 @@ import {
 } from '../settings/conversions'
 
 var schema = require('@signalk/signalk-schema');
-var signalkClient = require('@signalk/client').Client;
-var signalkUnits = new signalkClient;
 
 var SettingsPanel = (props) => {
   var analogSettings = (
@@ -75,44 +73,14 @@ require('util').inherits(Widget, BaseWidget);
 
 function Widget(id, options, streamBundle, instrumentPanel) {
   BaseWidget.call(this, id, options, streamBundle, instrumentPanel);
-  this.id = id;
-  this.streamBundle = streamBundle;
-  this.instrumentPanel = instrumentPanel;
-  this.options = options;
   this.options.selectedWidget = this.options.selectedWidget || 0;
   this.maxValueStream = new Bacon.Bus();
   this.minValueStream = new Bacon.Bus();
   this.redLineStream = new Bacon.Bus();
   var pathParts = options.path.split('.');
-  this.unit = '';
-  this.delayed = 0;
   this.widgets = [];
-//  this.unit = this.getUnitForPath(options.path) || "";
-  const schemaUnit = schema.getUnits('vessels.foo.' + options.path);
+  this.unit = this.getUnitForPath(options.path) || "";
 
-  if (schemaUnit === undefined) {
-    var widget = this;
-    widget.delayed = 1;
-    signalkUnits = new signalkClient(instrumentPanel.host.split(':')[0], instrumentPanel.host.split(':')[1], instrumentPanel.useSSL);
-    signalkUnits.apiGet('/vessels/self/' + options.path.replace(/\./g, '/') + '/meta/units')
-    .then(function (response) {
-      console.log(response.req.url + ':' + response.status + ':' + response.text);
-      widget.unit = JSON.parse(response.text);
-      widget.createUpdateWidgets(widget);
-      widget.delayed = 0;
-      widget.instrumentPanel.pushGridChanges();
-    })
-    .catch(function (error) {
-      console.log(error);
-      widget.createUpdateWidgets(widget);
-      widget.delayed = 0;
-      widget.instrumentPanel.pushGridChanges();
-    });
-  } else {
-       this.unit = schemaUnit;
-    }
-  this.createUpdateWidgets(this);    
-/*
   var valueStream = streamBundle.getStreamForSourcePath(options.sourceId, options.path);
   if ((options.convertTo) && (this.unit !== '')) {
     var conversions = getConversionsForUnit(this.unit);
@@ -122,7 +90,7 @@ function Widget(id, options, streamBundle, instrumentPanel) {
       console.error("No such conversion:" + this.unit + "=>" + options.convertTo)
     }
   }
-  
+
   this.widgets = [React.createElement(digitalComponent,{
     key: id,
     unit: this.unit,
@@ -147,59 +115,8 @@ function Widget(id, options, streamBundle, instrumentPanel) {
     redLine: options.redLine,
     redLineStream: this.redLineStream.toProperty()
   })];
-*/
-}
 
-Widget.prototype.createUpdateWidgets = function(widget) {
-  var valueStream = widget.streamBundle.getStreamForSourcePath(widget.options.sourceId, widget.options.path);
-
-  if (widget.unit === 'rad') {
-    widget.unit = 'deg';
-  }
-
-  if ((widget.options.convertTo) && (widget.unit !== '')) {
-    var conversions = getConversionsForUnit(widget.unit);
-    if (conversions[widget.options.convertTo]) {
-      valueStream = valueStream.map(conversions[widget.options.convertTo]);
-    } else {
-        console.error("No such conversion:" + widget.unit + "=>" + widget.options.convertTo)
-      }
-  }
-
-  if ((widget.delayed === 1) && (widget.widgets.length > 0)) {
-    var newWidgets  = [];
-    widget.widgets.forEach(function(item, index) {
-      newWidgets.push( React.cloneElement(item,{
-        unit: widget.unit,
-        valueStream: valueStream
-      }));
-    })
-    widget.widgets = newWidgets;
-	} else {
-      this.widgets = [React.createElement(digitalComponent,{
-        key: widget.id,
-        unit: widget.unit,
-        convertTo: widget.options.convertTo,
-        valueStream: valueStream,
-        label: widget.options.path,
-        path: widget.options.path,
-        sourceId: widget.options.sourceId,
-      }),React.createElement(analogComponent,{
-        key: widget.id,
-        unit: widget.unit,
-        convertTo: widget.options.convertTo,
-        valueStream: valueStream,
-        label: widget.options.path,
-        path: widget.options.path,
-        sourceId: widget.options.sourceId,
-        maxValue: widget.options.maxValue,
-        maxValueStream: widget.maxValueStream.toProperty(),
-        minValue: widget.options.minValue,
-        minValueStream: widget.minValueStream.toProperty(),
-        redLine: widget.options.redLine,
-        redLineStream: widget.redLineStream.toProperty()
-      })];
-    }
+  this.updateFromOnlineMeta(this);
 }
 
 Widget.prototype.getReactElement = function() {
