+++ lib/ui/widgets/windmeter.js
@@ -154,7 +154,6 @@ require('util').inherits(WindMeter, BaseWidget);
 
 function WindMeter(id, options, streamBundle, instrumentPanel) {
   BaseWidget.call(this, id, options, streamBundle, instrumentPanel);
-  this.options = options;
   this.options.activeSources = {};
   this.options.initialDimensions = this.options.initialDimensions || {h: 4};
   this.active = true;
